Split misalignment_qa into library and implementation layers.
Move framework code into aieng.agent_evals.misalignment_qa so it runs in CI,
keep bootcamp notebooks/configs in implementations/, and add evaluate.py as
the thin CLI entry point.
