+++ aieng-eval-agents/pyproject.toml
@@ -28,6 +28,7 @@ dependencies = [
     "httpx>=0.27.0",
     "tenacity>=8.2.0",
     "python-dotenv>=1.0.0",
+    "PyYAML>=6.0.0",
     "html-to-markdown>=2.24.0,<3.0.0",
 ]
